fix: rename evaluators rename to snakecase
same as #2962 (but avoiding merge conflicts)
fixes #1124
